Code passing lint errors
Code is now in right directory.
Bug fixex
