+++ src/main/python/gcloud_hydra/vm_files/add_grouping.sh
