+++ src/main/python/gcloud_hydra/vm_files/add_mesos_sphere_repo_and_install_java.sh
