+++ src/main/python/gcloud_hydra/vm_files/hydra_pkgs_install.sh
