+++ src/main/python/gcloud_hydra/vm_files/master_marathon_conf.sh
