+++ src/main/python/gcloud_hydra/vm_files/srv_init_rules_and_restart_srv.sh
